Separate out LdpConfig for reuse
